fix: remove initialization of Ohy, Osat, O33 and dz in Soil
The Ohy, Osat and O33 values were incorrectly initialized in the soil
parameters. In the original MATLAB code, they are created temporarily in
the function initializing the outputs. The updated implementation
mimicks this approach, by temporarily initializing these fields in the
water flux model variables, which are the only composite types using
these variables.

Issue(s): #98